requirements.txt
streamlit
python-nmap
ipaddress
regex
